+++ tutoraspects/__about__.py
@@ -2,4 +2,4 @@
 Expose some package metadata.
 """
 
-__version__ = "0.56.1"
+__version__ = "0.59.0"