+++ tutoraspects/templates/aspects/apps/superset/pythonpath/openedx_sso_security_manager.py
@@ -1,13 +1,12 @@
 import logging
 import time
-from functools import lru_cache
 
 import jwt
 import requests
 from authlib.common.urls import add_params_to_qs, add_params_to_uri
 from authlib.integrations.flask_client import OAuthError
 from flask import current_app, session
-from superset.constants import LRU_CACHE_MAX_SIZE
+from superset.extensions import cache_manager
 from superset.security import SupersetSecurityManager
 
 log = logging.getLogger(__name__)
@@ -69,17 +68,20 @@ class OpenEdxSsoSecurityManager(SupersetSecurityManager):
             oauth_remote = self.oauth_remotes.get(provider)
 
             response = oauth_remote.get(url).json()
-            language_preference = response.get("pref-lang", "en").replace("-", "_")[0:2]
+            locale_preference = response.get("pref-lang", "en").replace("-", "_")
 
-            if language_preference not in current_app.config["LANGUAGES"]:
+            if locale_preference not in current_app.config["DASHBOARD_LOCALES"]:
                 log.warning(
-                    f"Language {language_preference} is not supported by Superset"
+                    f"Language {locale_preference} is not supported by Superset"
                 )
-                language_preference = "en"
+                locale_preference = "en"
 
             user_roles = self._get_user_roles(
-                user_profile.get("preferred_username"), language_preference
+                user_profile.get("preferred_username"), locale_preference
             )
+
+            log.info(f"User roles for {user_profile['preferred_username']}:"
+                     f" {user_roles}")
 
             return {
                 "name": user_profile["name"],
@@ -156,32 +158,35 @@ class OpenEdxSsoSecurityManager(SupersetSecurityManager):
         """
         return self.get_oauth_token().get("access_token")
 
-    def _get_user_roles(self, username, language):
+    def _get_user_roles(self, username, locale):
         """
         Returns the Superset roles that should be associated with the given user.
         """
         decoded_access_token = self.decoded_user_info()
 
         if decoded_access_token.get("superuser", False):
-            return ["admin", "admin-{language}"]
+            return ["admin", f"admin-{locale}"]
         elif decoded_access_token.get("administrator", False):
-            return ["alpha", "operator", "operator-{language}"]
+            return ["alpha", "operator", f"operator-{locale}"]
         else:
-            # User has to have staff access to one or more courses to view any content here.
-            courses = self.get_courses(username)
+            # User has to have staff access to one or more courses to view any content
+            # here. Since this is only called on login, we take the opportunity
+            # to force refresh the user's courses from LMS. This allows them to bust
+            # the course permissions cache if necessary by logging out and back in.
+            courses = self.get_courses(username, force=True)
             if courses:
-                return ["instructor", f"instructor-{language}"]
+                return ["instructor", f"instructor-{locale}"]
             else:
                 roles = self.extra_get_user_roles(username, decoded_access_token)
                 if roles:
                     if {{SUPERSET_BLOCK_STUDENT_ACCESS}} and 'student' in roles:
                         raise Exception(f"Student access not allowed for {username} due to SUPERSET_BLOCK_STUDENT_ACCESS setting.")
                     return roles
-                
+
                 if {{SUPERSET_BLOCK_STUDENT_ACCESS}}:
                     raise Exception(f"Student {username} tried to access Superset")
                 else:
-                    return ["student", f"student-{language}"]
+                    return ["student", f"student-{locale}"]
 
     def extra_get_user_roles(self, username, decoded_access_token):
         """
@@ -190,10 +195,24 @@ class OpenEdxSsoSecurityManager(SupersetSecurityManager):
         {{patch("superset-sso-assignment-rules") | indent(8)}}
         return None
 
-    def get_courses(self, username, permission="staff", next_url=None):
+    def get_courses(self, username, permission="staff", next_url=None, force=False):
         """
         Returns the list of courses the current user has access to.
+
+        Calls itself recursively if the response is paginated.
+        Force will cause a force refresh of the cache, we do this on login to make
+        sure that the user always has an ability to update this data.
         """
+        cache = cache_manager.cache
+        cache_key = f"{username}+{permission}"
+
+        # Only return from the cache when we're at the top level
+        # (not in recursion)
+        if not next_url and not force:
+            obj = cache.get(cache_key)
+            if obj is not None:
+                return obj
+
         courses = []
         provider = session.get("oauth_provider")
         oauth_remote = self.oauth_remotes.get(provider)
@@ -211,7 +230,9 @@ class OpenEdxSsoSecurityManager(SupersetSecurityManager):
             username=username, permission=permission
         )
         url = next_url or courses_url
-        response = oauth_remote.get(url, token=token).json()
+        resp = oauth_remote.get(url, token=token)
+        resp.raise_for_status()
+        response = resp.json()
 
         for course in response.get("results", []):
             course_id = course.get("course_id")
@@ -225,5 +246,11 @@ class OpenEdxSsoSecurityManager(SupersetSecurityManager):
             )
             for course_id in next_courses:
                 courses.append(course_id)
+
+        log.info(f"Courses for {username}: {courses}")
+
+        # Only cache the top level (non-recursive) response
+        if not next_url:
+            cache.set(cache_key, courses, timeout=300)
 
         return courses